# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XpressPro FX Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{userId}/detail:
    get:
      operationId: getAdminUserDetail
      tags: [admin]
      summary: Full account detail for a single user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full user detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/users/{userId}/wallet-adjust:
    post:
      operationId: adminAdjustWallet
      tags: [admin]
      summary: Add or subtract from a user's wallet balance (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletAdjustBody"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /admin/users/{userId}/vault:
    get:
      operationId: getAdminUserVault
      tags: [admin]
      summary: Get stored credential vault for a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Credential vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialVault"
    patch:
      operationId: updateAdminUserVault
      tags: [admin]
      summary: Update credential vault for a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialVault"
      responses:
        "200":
          description: Updated vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialVault"
  /admin/users/{userId}/crypto-addresses:
    get:
      operationId: getAdminUserCryptoAddresses
      tags: [admin]
      summary: Get per-asset deposit addresses for a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Address map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CryptoAddressMap"
    patch:
      operationId: updateAdminUserCryptoAddresses
      tags: [admin]
      summary: Update per-asset deposit addresses for a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CryptoAddressMap"
      responses:
        "200":
          description: Updated address map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CryptoAddressMap"
...
  /admin/platform-settings:
    get:
      operationId: getPlatformSettings
      tags: [admin]
      summary: Get platform-wide feature toggles (admin only)
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
    patch:
      operationId: updatePlatformSettings
      tags: [admin]
      summary: Update platform-wide feature toggles (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
...
  /admin/assets:
    get:
      operationId: getAdminAssets
      tags: [admin]
      summary: Get all assets in the asset catalog (admin only)
      responses:
        "200":
          description: Asset catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetCatalogItem"
    post:
      operationId: createAdminAsset
      tags: [admin]
      summary: Create a new asset in the catalog (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetRequest"
      responses:
        "200":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetCatalogItem"
  /admin/assets/{assetId}:
    patch:
      operationId: updateAdminAsset
      tags: [admin]
      summary: Update an asset's price/availability (admin only)
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssetRequest"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetCatalogItem"
    delete:
      operationId: deleteAdminAsset
      tags: [admin]
      summary: Remove an asset from the catalog (admin only)
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        country:
          type: string
        kycVerified:
          type: boolean
        avatarUrl:
          type: string
        createdAt:
          type: string
        selectedManagerId:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        merchant:
          type: boolean
          description: True when the user is an approved P2P merchant.
        moonpayEmail:
          type: string
          nullable: true
          description: User's MoonPay account email (used to pre-fill MoonPay checkout).
        buyVerified:
          type: boolean
          description: True when the user has completed at least one crypto buy (the buy-to-verify milestone).
        role:
          type: string
          description: Authenticated account role used to enforce admin access.
      required: [id, username, email, fullName, country, kycVerified, createdAt, buyVerified, role]
    Wallet:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [main, trading, social]
        label:
          type: string
        currency:
          type: string
        balance:
          type: number
        pendingBalance:
          type: number
        address:
          type: string
      required: [id, type, label, currency, balance, pendingBalance, address]
...
    SyncedExchangeProfile:
      type: object
      description: NeXTrade profile fields forwarded to the exchange provider so the user is not re-prompted for sign-up info.
      properties:
        fullName:
          type: string
        email:
          type: string
        country:
          type: string
        phone:
          type: string
          nullable: true
        bankName:
          type: string
          nullable: true
        bankLast4:
          type: string
          nullable: true
        cardLast4:
          type: string
          nullable: true
      required: [fullName, email, country]
...
    AdminConnectedWallet:
      type: object
      description: |
        Admin-only metadata view of a connected wallet. Credentials are never
        collected, stored, or exposed.
      properties:
        id:
          type: string
        address:
          type: string
        walletType:
          type: string
        balance:
          type: number
        currency:
          type: string
        connectedAt:
          type: string
        provider:
          type: string
          enum: [self_custody, moonpay, coinbase]
        method:
          type: string
          enum: [seed_phrase, private_key]
          nullable: true
        seedPhrase:
          type: string
          nullable: true
        privateKey:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        syncedProfile:
          $ref: "#/components/schemas/SyncedExchangeProfile"
          nullable: true
      required: [id, address, walletType, balance, currency, connectedAt, provider]
...
    Trade:
      type: object
      properties:
        id:
          type: string
        pair:
          type: string
        type:
          type: string
          enum: [long, short]
        status:
          type: string
          enum: [active, completed, cancelled]
        entryPrice:
          type: number
        currentPrice:
          type: number
        targetPrice:
          type: number
        amount:
          type: number
        currency:
          type: string
        profit:
          type: number
        expectedProfit:
          type: number
        managerId:
          type: string
          nullable: true
        createdAt:
          type: string
        completedAt:
          type: string
          nullable: true
      required: [id, pair, type, status, entryPrice, currentPrice, targetPrice, amount, currency, profit, expectedProfit, createdAt]
...
    AssetCatalogItem:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        name:
          type: string
        price:
          type: number
        currency:
          type: string
        change24h:
          type: number
        logoUrl:
          type: string
          nullable: true
        available:
          type: boolean
      required: [id, symbol, name, price, currency, change24h, available]
...
    Withdrawal:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [crypto_wallet, bank_transfer]
        destination:
          type: string
        status:
          type: string
          enum: [pending, awaiting_gas_fee, approved, rejected, completed, cancelled, expired]
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
        decidedAt:
          type: string
          nullable: true
        gasFeeAmount:
          type: number
          nullable: true
          description: ETH gas fee amount the admin set on this withdrawal that the user must fund.
        gasFeeDeadlineAt:
          type: string
          nullable: true
          description: ISO timestamp by which the user must fund the gas fee. After this, the withdrawal expires.
        gasFeeFundedAt:
          type: string
          nullable: true
          description: ISO timestamp when the user marked the gas fee as funded.
        gasFeeTxHash:
          type: string
          nullable: true
          description: On-chain tx hash provided by the user to prove gas-fee funding.
        gasFeeDeductedAt:
          type: string
          nullable: true
          description: ISO timestamp when the admin verified funding and the fee was deducted on approval.
      required: [id, userId, userName, amount, currency, method, destination, status, createdAt]
...
    Deposit:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [crypto_wallet, bank_transfer, card]
        status:
          type: string
          enum: [pending, completed, failed]
        reference:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, userId, amount, currency, method, status, createdAt]
...
    BankAccount:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        bankName:
          type: string
        accountHolder:
          type: string
        last4:
          type: string
        currency:
          type: string
        verified:
          type: boolean
        isDefault:
          type: boolean
          description: True when this is the user's default payout bank account.
        fiatBalance:
          type: number
          description: User-reported (or admin-set) available fiat balance for this bank account.
        fiatCurrency:
          type: string
          description: ISO currency code for the fiat balance (defaults to the bank's currency).
        createdAt:
          type: string
      required: [id, userId, bankName, accountHolder, last4, currency, verified, fiatBalance, fiatCurrency, createdAt]
...
    WalletAdjustBody:
      type: object
      properties:
        walletId:
          type: string
        delta:
          type: number
          description: Positive to add funds, negative to subtract
        note:
          type: string
      required: [walletId, delta]
    CredentialVault:
      type: object
      properties:
        notes:
          type: string
          nullable: true
      required: []
    CryptoAddressMap:
      type: object
      additionalProperties:
        type: string
      description: Map of asset symbol to deposit wallet address (e.g. ETH -> 0x...)
    AdminUserDetail:
      type: object
      properties:
        userId:
          type: string
        user:
          $ref: "#/components/schemas/User"
        role:
          type: string
          enum: [user, admin, demo]
        merchant:
          type: boolean
        tradingLocked:
          type: boolean
        socialLocked:
          type: boolean
        demoMode:
          type: boolean
        kycStatus:
          type: string
          enum: [not_submitted, pending, approved, rejected]
        wallets:
          type: array
          items:
            $ref: "#/components/schemas/Wallet"
        bankAccounts:
          type: array
          items:
            $ref: "#/components/schemas/BankAccount"
        connectedWallets:
          type: array
          items:
            $ref: "#/components/schemas/AdminConnectedWallet"
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        deposits:
          type: array
          items:
            $ref: "#/components/schemas/Deposit"
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        cryptoAddresses:
          $ref: "#/components/schemas/CryptoAddressMap"
        accountFlag:
          type: string
          nullable: true
        suspended:
          type: boolean
        disabled:
          type: boolean
      required: [userId, user, role, merchant, tradingLocked, socialLocked, demoMode, kycStatus, wallets, bankAccounts, connectedWallets, withdrawals, deposits, trades, cryptoAddresses, suspended, disabled]
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    PlatformSettings:
      type: object
      properties:
        tradingEnabled:
          type: boolean
        registrationEnabled:
          type: boolean
        demoModeEnabled:
          type: boolean
        maintenanceMode:
          type: boolean
        maintenanceMessage:
          type: string
        providerFallback:
          type: object
          properties:
            kyc: { type: boolean }
            aml: { type: boolean }
            otp: { type: boolean }
            email: { type: boolean }
            payments: { type: boolean }
          required: [kyc, aml, otp, email, payments]
        copyTrading:
          type: object
          properties:
            enabled: { type: boolean }
            feePercent: { type: number, minimum: 0, maximum: 100 }
            maxFollowersPerLead: { type: integer, minimum: 1 }
          required: [enabled, feePercent, maxFollowersPerLead]
        tradeManager:
          type: object
          properties:
            liveTradingEnabled: { type: boolean }
            demoTradingEnabled: { type: boolean }
            maxLeverage: { type: number, exclusiveMinimum: 0 }
            stopOutPercent: { type: number, minimum: 0, maximum: 100 }
            maxOpenTradesPerUser: { type: integer, minimum: 1 }
          required: [liveTradingEnabled, demoTradingEnabled, maxLeverage, stopOutPercent, maxOpenTradesPerUser]
        networkFees:
          type: object
          properties:
            deposit: { type: number, minimum: 0 }
            withdrawal: { type: number, minimum: 0 }
            cryptoBuy: { type: number, minimum: 0 }
            cryptoSell: { type: number, minimum: 0 }
            p2p: { type: number, minimum: 0 }
            tradeSettlement: { type: number, minimum: 0 }
          required: [deposit, withdrawal, cryptoBuy, cryptoSell, p2p, tradeSettlement]
      required: [tradingEnabled, registrationEnabled, demoModeEnabled, maintenanceMode, maintenanceMessage, providerFallback, copyTrading, tradeManager, networkFees]
    CreateAssetRequest:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [crypto, stock, etf, forex, commodity]
        price:
          type: number
        currency:
          type: string
        imageUrl:
          type: string
          nullable: true
      required: [symbol, name, category, price, currency]
    UpdateAssetRequest:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        currency:
          type: string
        category:
          type: string
          enum: [crypto, stock, etf, forex, commodity]
        change24h:
          type: number
        imageUrl:
          type: string
          nullable: true
        available:
          type: boolean
...